requirements.txt
openai>=1.0.0
pandas
openpyxl
pymupdf
python-dotenv
python-docx
